Improve everything:
The sync thread runs in the background and periodically checks for changes in the DNS records. It uses a backoff strategy to avoid overwhelming the DNS servers with too many requests.

The users have rights to modify their own DNS records and the zones belonging to them.
There is a group ownership as well.

Docs were improved, Dockerfile added, I guess it's time for testing.